feat(patient-personal-log): page which displays a table of the patient's history of uploads  (#21)
A page which can be accessed by the patient, from the menu list on the navbar by selecting "History". This page displays a table containing all the patients previous uploads. The uploads can be sorted based on date. Each row has an option to "view" the upload. This needs to be changed later on, once the page which displays the details is created, for now it links to the home page. The data displayed uses mock data, this will later be updated to use the data from the database.