Fix test
The test was broken as there was a time difference between the time stamp stored in the database and the time
stamp expectd due to the small delay of saving to the database. The difference is very small, typically much
less than even milliseconds. Hence, the exact time was ignored and was replaced by checking that the time does
in fact exist.